+++ Sources/TuistKit/Commands/TrackableCommand/TrackableParametersDelegate.swift
@@ -7,6 +7,6 @@ protocol HasTrackableParameters {
 
 /// `TrackableParametersDelegate` contains the callback that should be called
 /// before running a command, with extra parameters that are only known at runtime
-protocol TrackableParametersDelegate {
+protocol TrackableParametersDelegate: AnyObject {
     func willRun(withParameters: [String: String])
 }