+++ Sources/TuistKit/ProjectEditor/ProjectEditorMapper.swift
@@ -23,7 +23,9 @@ protocol ProjectEditorMapping: AnyObject {
     ) throws -> ValueGraph
 }
 
+// swiftlint:disable:next type_body_length
 final class ProjectEditorMapper: ProjectEditorMapping {
+    // swiftlint:disable:next function_body_length
     func map(
         name: String,
         tuistPath: AbsolutePath,
@@ -116,6 +118,7 @@ final class ProjectEditorMapper: ProjectEditorMapping {
         )
     }
 
+    // swiftlint:disable:next function_body_length
     private func mapManifestsProject(
         projectManifests: [AbsolutePath],
         projectDescriptionPath: AbsolutePath,
@@ -250,6 +253,7 @@ final class ProjectEditorMapper: ProjectEditorMapping {
         )
     }
 
+    // swiftlint:disable:next function_body_length
     private func mapPluginsProject(
         pluginManifests: [EditablePluginManifest],
         projectDescriptionPath: AbsolutePath,
@@ -272,7 +276,10 @@ final class ProjectEditorMapper: ProjectEditorMapping {
         let pluginTargets = pluginManifests.map { manifest -> Target in
             let pluginManifest = manifest.path.appending(component: "Plugin.swift")
             let pluginHelpersPath = manifest.path.appending(component: Constants.helpersDirectoryName)
-            let sourcePaths = [pluginManifest] + FileHandler.shared.glob(pluginHelpersPath, glob: "**/*.swift")
+            let pluginTemplatesPath = manifest.path.appending(component: Constants.templatesDirectoryName)
+            let sourcePaths = [pluginManifest] +
+                FileHandler.shared.glob(pluginHelpersPath, glob: "**/*.swift") +
+                FileHandler.shared.glob(pluginTemplatesPath, glob: "**/*.swift")
             return editorHelperTarget(
                 name: manifest.name,
                 filesGroup: pluginsFilesGroup,