+++ Sources/TuistKit/Services/Lint/LintCodeService.swift
@@ -9,8 +9,6 @@ import TuistLoader
 import TuistSupport
 
 enum LintCodeServiceError: FatalError, Equatable {
-    /// Thrown when neither a workspace or a project is found in the given path.
-    case manifestNotFound(AbsolutePath)
     /// Thrown when target with given name does not exist.
     case targetNotFound(String)
     /// Throws when no lintable files found for target with given name.
@@ -19,7 +17,7 @@ enum LintCodeServiceError: FatalError, Equatable {
     /// Error type.
     var type: ErrorType {
         switch self {
-        case .manifestNotFound, .targetNotFound, .lintableFilesForTargetNotFound:
+        case .targetNotFound, .lintableFilesForTargetNotFound:
             return .abort
         }
     }
@@ -27,8 +25,6 @@ enum LintCodeServiceError: FatalError, Equatable {
     /// Description
     var description: String {
         switch self {
-        case let .manifestNotFound(path):
-            return "Couldn't find Project.swift nor Workspace.swift at \(path.pathString)"
         case let .targetNotFound(name):
             return "Target with name '\(name)' not found in the project."
         case let .lintableFilesForTargetNotFound(name):
@@ -38,26 +34,26 @@ enum LintCodeServiceError: FatalError, Equatable {
 }
 
 final class LintCodeService {
-    private let rootDirectoryLocator: RootDirectoryLocating
     private let codeLinter: CodeLinting
-    private let graphLoader: ValueGraphLoading
-    private let modelLoader: GeneratorModelLoading
-    private let manifestLoading: ManifestLoading
+    private let manifestGraphLoader: ManifestGraphLoading
 
-    init(rootDirectoryLocator: RootDirectoryLocating = RootDirectoryLocator(),
-         codeLinter: CodeLinting = CodeLinter(),
-         manifestLoading: ManifestLoading = ManifestLoader(),
-         modelLoader: GeneratorModelLoading = GeneratorModelLoader(
-             manifestLoader: ManifestLoader(),
-             manifestLinter: AnyManifestLinter()
-         ),
-         graphLoader: ValueGraphLoading = ValueGraphLoader())
-    {
-        self.rootDirectoryLocator = rootDirectoryLocator
+    convenience init() {
+        let manifestLoader = ManifestLoaderFactory()
+            .createManifestLoader()
+        let manifestGraphLoader = ManifestGraphLoader(manifestLoader: manifestLoader)
+        let codeLinter = CodeLinter()
+        self.init(
+            codeLinter: codeLinter,
+            manifestGraphLoader: manifestGraphLoader
+        )
+    }
+
+    init(
+        codeLinter: CodeLinting,
+        manifestGraphLoader: ManifestGraphLoading
+    ) {
         self.codeLinter = codeLinter
-        self.manifestLoading = manifestLoading
-        self.modelLoader = modelLoader
-        self.graphLoader = graphLoader
+        self.manifestGraphLoader = manifestGraphLoader
     }
 
     func run(path: String?, targetName: String?, strict: Bool) throws {
@@ -65,7 +61,8 @@ final class LintCodeService {
         let path = self.path(path)
 
         // Load graph
-        let graph = try loadDependencyGraph(at: path)
+        logger.notice("Loading the dependency graph at \(path)")
+        let graph = try manifestGraphLoader.loadGraph(at: path)
 
         // Get sources
         let graphTraverser = ValueGraphTraverser(graph: graph)
@@ -82,28 +79,6 @@ final class LintCodeService {
         guard let path = path else { return FileHandler.shared.currentPath }
 
         return AbsolutePath(path, relativeTo: FileHandler.shared.currentPath)
-    }
-
-    // MARK: - Load dependency graph
-
-    private func loadDependencyGraph(at path: AbsolutePath) throws -> ValueGraph {
-        let manifests = manifestLoading.manifests(at: path)
-
-        logger.notice("Loading the dependency graph")
-        if manifests.contains(.workspace) {
-            logger.notice("Loading workspace at \(path.pathString)")
-            let workspace = try modelLoader.loadWorkspace(at: path)
-            let projects = try workspace.projects.map(modelLoader.loadProject)
-            let graph = try graphLoader.loadWorkspace(workspace: workspace, projects: projects)
-            return graph
-        } else if manifests.contains(.project) {
-            logger.notice("Loading project at \(path.pathString)")
-            let project = try modelLoader.loadProject(at: path)
-            let (_, graph) = try graphLoader.loadProject(at: path, projects: [project])
-            return graph
-        } else {
-            throw LintCodeServiceError.manifestNotFound(path)
-        }
     }
 
     // MARK: - Get sources to lint