+++ Sources/TuistLoader/Up/UpHomebrew.swift
@@ -47,7 +47,8 @@ class UpHomebrew: Up, GraphInitiatable {
         if !toolInstalled("brew") {
             logger.notice("Installing Homebrew")
             try System.shared.runAndPrint(
-                "/usr/bin/ruby",
+                "/usr/bin/env",
+                "ruby",
                 "-e",
                 "\"$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)\"",
                 verbose: true,
@@ -58,7 +59,8 @@ class UpHomebrew: Up, GraphInitiatable {
         try nonInstalledPackages.forEach { package in
             logger.notice("Installing Homebrew package: \(package)")
             try System.shared.runAndPrint(
-                "/usr/local/bin/brew",
+                "/usr/bin/env",
+                "brew",
                 "install",
                 package,
                 verbose: true,
@@ -68,6 +70,6 @@ class UpHomebrew: Up, GraphInitiatable {
     }
 
     private func packageInstalled(_ name: String) -> Bool {
-        (try? System.shared.run("/usr/local/bin/brew", "list", name)) != nil
+        (try? System.shared.run("/usr/bin/env", "brew", "list", name)) != nil
     }
 }