+++ Tests/Fixtures/WorkspaceWithPlugins/Tuist/Config.swift
@@ -1,0 +1,8 @@
+import ProjectDescription
+
+let config = Config(
+    plugins: [
+        .local(path: "//LocalPlugin"),
+    ],
+    generationOptions: []
+)