+++ Tests/Fixtures/WorkspaceWithPlugins/Workspace.swift
@@ -1,0 +1,9 @@
+import LocalPlugin
+import ProjectDescription
+
+let workspace = Workspace(
+    name: "Workspace",
+    projects: [
+        "App",
+    ]
+)