+++ projects/fourier/test/test_helper.rb
@@ -10,6 +10,8 @@ require "tmpdir"
 require "tempfile"
 require "byebug"
 
+Dir.glob(File.join(__dir__, "test_helpers/*")).each { |f| require(f) }
+
 CLI::UI::StdoutRouter.enable
 
 require "minitest/autorun"